requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
extension==1.0.0
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
Markdown==3.3.7
MarkupSafe==2.1.0
mysql-connector-python==8.0.28
protobuf==3.19.4
pycodestyle==2.8.0
pyshorteners==1.0.1
python-dotenv==0.20.0
requests==2.27.1
SQLAlchemy==1.4.31
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.8
Werkzeug==2.0.3
